refactor(app): rename Languages to ProjectInfo
