+++ src/app/App.svelte
@@ -1,7 +1,7 @@
 <script>
   import { Router, Route, Link } from "svelte-routing";
   import LandingPage from "./pages/LandingPage.svelte";
-  import Languages from "./pages/Languages.svelte";
+  import ProjectInfo from "./pages/ProjectInfo.svelte";
   import routes from "./routes";
 
   // Used for SSR. A falsy value is ignored by the Router.
@@ -10,5 +10,5 @@
 
 <Router {url}>
   <Route path={routes.landingPage} component={LandingPage} />
-  <Route path={routes.projectInfo} component={Languages} />
+  <Route path={routes.projectInfo} component={ProjectInfo} />
 </Router>
