+++ src/app/pages/ProjectInfo.svelte
